collision avoidanceを導入

+++ lidar_system/include/degree2position.hpp
@@ -5,7 +5,7 @@
 
 cv::Point2f degree2position(int x, int y, float angle, float distance) {
   float radian = angle * CV_PI / 180.0f;
-  return cv::Point2f(x + distance * std::cos(radian),
+  return cv::Point2d(x + distance * std::cos(radian),
                      y + distance * std::sin(radian));
 }
